+++ libnethack/src/pickup.c
@@ -1,5 +1,5 @@
 /* vim:set cin ft=c sw=4 sts=4 ts=8 et ai cino=Ls\:0t0(0 : -*- mode:c;fill-column:80;tab-width:8;c-basic-offset:4;indent-tabs-mode:nil;c-file-style:"k&r" -*-*/
-/* Last modified by Alex Smith, 2013-12-26 */
+/* Last modified by Alex Smith, 2013-12-29 */
 /* Copyright (c) Stichting Mathematisch Centrum, Amsterdam, 1985. */
 /* NetHack may be freely redistributed.  See license for details. */
 
@@ -1016,7 +1016,8 @@ pickup_object(struct obj *obj, long count, boolean telekinesis)
     int res, nearload;
 
     if (obj->quan < count) {
-        impossible("pickup_object: count %ld > quan %ld?", count, obj->quan);
+        impossible("pickup_object: count %ld > quan %ld?",
+                   (long)count, (long)obj->quan);
         return 0;
     }
